+++ packages/editor/editor-core/examples/1-labs-async-collapsed-editor.tsx
@@ -1,0 +1,30 @@
+import * as React from 'react';
+import CollapsedEditor from './../src/labs/CollapsedEditor';
+
+export interface State {
+  isExpanded: boolean;
+}
+export default class Example extends React.Component<{}, State> {
+  state = { isExpanded: false };
+
+  toggleExpanded = () => {
+    this.setState(prevState => ({ isExpanded: !prevState.isExpanded }));
+  };
+
+  render() {
+    return (
+      <CollapsedEditor
+        placeholder="What would you like to say?"
+        isExpanded={this.state.isExpanded}
+        onClickToExpand={this.toggleExpanded}
+        renderEditor={Editor => (
+          <Editor
+            appearance="comment"
+            onSave={() => alert('Saved!')}
+            onCancel={this.toggleExpanded}
+          />
+        )}
+      />
+    );
+  }
+}