+++ packages/editor/editor-core/src/ui/Appearance/FullPage.tsx
@@ -16,6 +16,7 @@ const FullPageEditorWrapper = styled.div`
   height: 100%;
   display: flex;
   flex-direction: column;
+  padding-bottom: 55px;
 `;
 FullPageEditorWrapper.displayName = 'FullPageEditorWrapper';
 
@@ -44,7 +45,6 @@ const ContentArea = styled.div`
   & .ProseMirror {
     flex-grow: 1;
     box-sizing: border-box;
-    padding-bottom: 55px;
   }
 
   && .ProseMirror {