+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/renderer
 
+## 13.0.17
+
+## 13.0.16
+- [patch] Move types/interfaces for ExtensionHandlers to editor-common [3d26cab](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3d26cab)
+
+## 13.0.15
+
+## 13.0.14
+- [patch] Prevent CodeBlocks from overflowing their container [50cc975](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/50cc975)
+
+## 13.0.13
+- [patch] Upgrading ProseMirror Libs [35d14d5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/35d14d5)
+
 ## 13.0.12
 
 ## 13.0.11