+++ packages/elements/emoji/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/emoji
 
+## 32.0.0
+- [major] FS-1697 move elements packages to use util-data-test for test data [deb820a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/deb820a)
+
 ## 31.1.23
 - [patch] Fix painting issue after scrolling to top of emoji list [debbe10](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/debbe10)